+++ samples/expo/app/(tabs)/index.tsx
@@ -38,7 +38,6 @@ Sentry.init({
         // default: [/.*/]
         failedRequestTargets: [/.*/],
       }),
-      Sentry.Integrations.Spotlight(),
     );
     return integrations.filter(i => i.name !== 'Dedupe');
   },
@@ -63,6 +62,7 @@ Sentry.init({
   _experiments: {
     profilesSampleRate: 0,
   },
+  enableSpotlight: true,
 });
 
 export default function TabOneScreen() {