+++ scripts/expo-upload-sourcemaps.js
@@ -4,8 +4,8 @@ const fs = require('fs');
 const path = require('path');
 const process = require('process');
 
+const SENTRY_ORG = 'SENTRY_ORG';
 const SENTRY_PROJECT = 'SENTRY_PROJECT';
-// The sentry org is inferred from the auth token
 const SENTRY_AUTH_TOKEN = 'SENTRY_AUTH_TOKEN';
 const SENTRY_CLI_EXECUTABLE = 'SENTRY_CLI_EXECUTABLE';
 
@@ -101,25 +101,42 @@ function groupAssets(assetPaths) {
   return groups;
 }
 
+let sentryOrg = getEnvVar(SENTRY_ORG);
 let sentryProject = getEnvVar(SENTRY_PROJECT);
 let authToken = getEnvVar(SENTRY_AUTH_TOKEN);
 const sentryCliBin = getEnvVar(SENTRY_CLI_EXECUTABLE) || require.resolve('@sentry/cli/bin/sentry-cli');
 
-if (!sentryProject) {
-  console.log(`🐕 Fetching ${SENTRY_PROJECT} from expo config...`);
+if (!sentryOrg || !sentryProject) {
+  console.log('🐕 Fetching from expo config...');
   const pluginConfig = getSentryPluginPropertiesFromExpoConfig();
   if (!pluginConfig) {
     console.error("Could not fetch '@sentry/react-native' plugin properties from expo config.");
     process.exit(1);
   }
-  if (!pluginConfig.project) {
-    console.error(
-      `Could not resolve sentry project, set it in the environment variable ${SENTRY_PROJECT} or in the '@sentry/react-native' plugin properties in your expo config.`,
-    );
-    process.exit(1);
+
+  if (!sentryOrg) {
+    if (!pluginConfig.organization) {
+      console.error(
+        `Could not resolve sentry org, set it in the environment variable ${SENTRY_ORG} or in the '@sentry/react-native' plugin properties in your expo config.`,
+      );
+      process.exit(1);
+    }
+
+    sentryOrg = pluginConfig.organization;
+    console.log(`${SENTRY_ORG} resolved to ${sentryOrg} from expo config.`);
   }
-  sentryProject = pluginConfig.project;
-  console.log(`${SENTRY_PROJECT} resolved to ${sentryProject} from expo config.`);
+
+  if (!sentryProject) {
+    if (!pluginConfig.project) {
+      console.error(
+        `Could not resolve sentry project, set it in the environment variable ${SENTRY_PROJECT} or in the '@sentry/react-native' plugin properties in your expo config.`,
+      );
+      process.exit(1);
+    }
+
+    sentryProject = pluginConfig.project;
+    console.log(`${SENTRY_PROJECT} resolved to ${sentryProject} from expo config.`);
+  }
 }
 
 if (!authToken) {
@@ -137,6 +154,8 @@ if (!outputDir) {
 const files = getAssetPathsSync(outputDir);
 const groupedAssets = groupAssets(files);
 
+const totalAssets = Object.keys(groupedAssets).length;
+let numAssetsUploaded = 0;
 for (const [assetGroupName, assets] of Object.entries(groupedAssets)) {
   const sourceMapPath = assets.find(asset => asset.endsWith('.map'));
   if (sourceMapPath) {
@@ -145,16 +164,30 @@ for (const [assetGroupName, assets] of Object.entries(groupedAssets)) {
       sourceMap.debug_id = sourceMap.debugId;
     }
     writeJSONFile(sourceMapPath, sourceMap);
+    console.log(`⬆️ Uploading ${assetGroupName} bundle and sourcemap...`);
+  } else {
+    console.log(`❓ Sourcemap for ${assetGroupName} not found, skipping...`);
+    continue;
   }
-  console.log(`⬆️ Uploading ${assetGroupName} bundle and sourcemap...`);
+
   const isHermes = assets.find(asset => asset.endsWith('.hbc'));
   execSync(`${sentryCliBin} sourcemaps upload ${isHermes ? '--debug-id-reference' : ''} ${assets.join(' ')}`, {
     env: {
       ...process.env,
       [SENTRY_PROJECT]: sentryProject,
+      [SENTRY_ORG]: sentryOrg,
     },
     stdio: 'inherit',
   });
+  numAssetsUploaded++;
 }
 
-console.log('✅ Uploaded bundles and sourcemaps to Sentry successfully.');
+if (numAssetsUploaded === totalAssets) {
+  console.log('✅ Uploaded bundles and sourcemaps to Sentry successfully.');
+} else {
+  console.warn(
+    `⚠️  Uploaded ${numAssetsUploaded} of ${totalAssets} bundles and sourcemaps. ${
+      numAssetsUploaded === 0 ? 'Ensure you are running `expo export` with the `--dump-sourcemap` flag.' : ''
+    }`,
+  );
+}